+++ src/controllers/ConnectionController.php
@@ -1,16 +1,29 @@
 <?php
+
 namespace imarc\googlecustomsearch\controllers;
 
-use Craft;
 use craft\web\Controller;
 use imarc\googlecustomsearch\Plugin;
 
-/**
- *
- */
 class ConnectionController extends Controller
 {
-    public function actionTest()
+    /**
+     * @var array|bool|int
+     */
+    public $allowAnonymous = false;
+
+    public function beforeAction($action): bool
+    {
+        if (!parent::beforeAction($action)) {
+            return false;
+        }
+
+        $this->requireAdmin();
+
+        return true;
+    }
+
+    public function actionTest(): \yii\web\Response
     {
         return $this->asJson(Plugin::getInstance()->search->testConnection());
     }